requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
lightgbm==4.6.0
MLB-StatsAPI==1.9.0
numpy==2.4.2
oauthlib==3.3.1
openpyxl==3.1.5
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
